+++ html/html.go
@@ -1,6 +1,6 @@
 /*
 	rulehunter - A server to find rules in data based on user specified goals
-	Copyright (C) 2016 vLife Systems Ltd <http://vlifesystems.com>
+	Copyright (C) 2016-2017 vLife Systems Ltd <http://vlifesystems.com>
 
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU Affero General Public License as published by
@@ -88,6 +88,9 @@ func generate(
 		if err := generateTagPages(config); err != nil {
 			return err
 		}
+		if err := generateCategoryPages(config); err != nil {
+			return err
+		}
 		if err := generateActivityPage(config, pm); err != nil {
 			return err
 		}
@@ -99,6 +102,9 @@ func generate(
 			return err
 		}
 		if err := generateTagPages(config); err != nil {
+			return err
+		}
+		if err := generateCategoryPages(config); err != nil {
 			return err
 		}
 		if err := generateActivityPage(config, pm); err != nil {
@@ -170,6 +176,7 @@ func makeHtmlNav(menuItem string) template.HTML {
 		"licence",
 		"reports",
 		"tag",
+		"category",
 		"activity",
 	}