+++ html/nav_tpl.go
@@ -1,6 +1,6 @@
 /*
 	rulehunter - A server to find rules in data based on user specified goals
-	Copyright (C) 2016 vLife Systems Ltd <http://vlifesystems.com>
+	Copyright (C) 2016-2017 vLife Systems Ltd <http://vlifesystems.com>
 
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU Affero General Public License as published by
@@ -40,6 +40,10 @@ const navTpl = `
 					<li class="active"><a href="">Reports</a></li>
 				{{else}}
 					<li><a href="">Reports</a></li>
+				{{end}}
+
+				{{if eq .MenuItem "category"}}
+					<li class="active"><a href=".">Category</a></li>
 				{{end}}
 
 				{{if eq .MenuItem "tag"}}